% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo-package.R
\docType{package}
\name{demo-package}
\alias{demo}
\alias{demo-package}
\title{demo: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
